Add validate single file to __main__
